moved hypothesisutils out of rrid.py and into its own file since other code needs to call on it as well
